corrections

+++ src/main/java/com/personalproject/inventorymanagementsystem/model/ProductPerformance.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class ProductPerformance {
     private Integer productId;
     private String productName;
-    private Long totalSales;
+    private BigDecimal totalSales;
 }
